Reorganized folder structure; Removed MultipleChoiceSurvey
