+++ resources/js/components/module/data/moduleData.js
