+++ resources/js/components/module/data/modules/alcoholModule.js
