+++ resources/js/components/module/data/modules/comparisonModule.js
