+++ resources/js/components/module/data/modules/marijuanaModule.js
