+++ resources/js/components/module/data/modules/nicotineModule.js
