+++ resources/js/components/module/data/modules/prescriptionDrugsModule.js
