+++ resources/js/components/module/shared/ModuleFooter.vue
