+++ resources/js/components/module/shared/ModuleProgressBar.vue
