+++ resources/js/components/module/templates/CardFlipTemplate.vue
