+++ resources/js/components/module/templates/EmailForm.vue
