+++ resources/js/components/module/templates/FinalSlide.vue
