+++ resources/js/components/module/templates/InfoTemplate.vue
