+++ resources/js/components/module/templates/InputPieChartTemplate.vue
