+++ resources/js/components/module/templates/IntroSlide.vue
