+++ resources/js/components/module/templates/QuizTemplate.vue
