+++ resources/js/components/module/templates/VideoTemplate.vue
