+++ resources/js/components/module/templates/info/InfoCarousel.vue
