+++ resources/js/components/module/templates/info/InfoList.vue
