+++ resources/js/components/module/templates/info/InfoParagraph.vue
