+++ resources/js/components/module/templates/info/InfoPhoto.vue
