+++ resources/js/components/module/templates/info/InfoPhotoList.vue
