+++ resources/js/components/module/templates/info/infoPhotoGrid.vue
