+++ resources/js/components/module/templates/input/InputPieChartLegend.vue
