--- resources/js/components/modules/templates/MultiChoiceSurveyResults.vue
@@ -1,60 +1,0 @@
-<template>
-  <div class="multiChoiceSurveyResults">
-    <div v-if="surveyResultsEmailWasSubmitted===true" class="alert alert-success transition-350ms">
-      <strong>Success!</strong> Email was successfully sent
-    </div>
-    <div v-if="surveyResultsEmailWasSubmitted===false" class="alert alert-danger transition-350ms">
-      <strong>Failure. </strong> An error has occured
-    </div>
-
-    <h2 class="multiChoiceSurveyResults__heading">
-      Based on your 3 answers here is tailored motivational and/or educational
-      advice to follow.
-      These are also resources to check out based off your 3 answers.
-    </h2>
-    <div class="multiChoiceSurveyResults__link-container">
-      <ul>
-        <li v-for="(url, index) in surveyResultLinks" :key="index" class="multiChoiceSurveyResults__link" >
-            <a :href="url" target="_blank"> Visit this example link</a>
-        </li>
-      </ul>
-    </div>
-    <h2 @click="sendUserSurveyResults" class="multiChoiceSurveyResults__mail">
-      <i class="fas fa-envelope"></i>
-      Email these results to yourself
-    </h2>
-  </div>
-
-</template>
-<script>
-import { mapActions, mapGetters } from 'vuex'
-import { awaitTimeBeforeContinue } from './../../../mixins/awaitTimeBeforeContinue'
-
-export default {
-  mixins: [awaitTimeBeforeContinue],
-  computed: {
-    ...mapGetters(
-      [
-        'surveyResultLinks',
-        'surveyResultsEmailWasSubmitted'
-      ]
-    )
-  },
-
-  methods: {
-    ...mapActions(
-      [
-        'emailSurveyResults'
-      ]
-    ),
-    sendUserSurveyResults () {
-      var links = '<ul>'
-      for (var i = 0; i < this.surveyResultLinks.length; i++) {
-        links += (`<li><a href=${this.surveyResultLinks[i]} target=_blank>${this.surveyResultLinks[i]}</a></li>`)
-      }
-      links += ('</ul>')
-      this.emailSurveyResults(links)
-    }
-  }
-}
-</script>
